dd-1052162 Add Network Security Group (#16)
* dd-1052162 formatting

* dd-1052162 Add Network Security Group

Thanks to Jianguo Ma, I know that the absense of an NSG from this template caused Azure to
apply an auto-generated one.  This one is too strong.

By creating our own, we can prevent the network problems that prevent access.

There is one additional step that must be performed after deployment to the Oracle Enterprise Java
subscriptin, however.

* Security rules to delete.

   * 103 NRMS-Rule-103

   * 105 NRMS-Rule-105

* dd-1052162 Add Network Security Group

Thanks to Jianguo Ma, I know that the absense of an NSG from this template caused Azure to
apply an auto-generated one.  This one is too strong.

By creating our own, we can prevent the network problems that prevent access.

There is one additional step that must be performed after deployment to the Oracle Enterprise Java
subscriptin, however.

* Security rules to delete.

   * 103 NRMS-Rule-103

   * 105 NRMS-Rule-105